+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   @lumx/core: improved color contrast on dark L6 & L2
 
+### Added
+
+-   New `Navigation` component.
+
 ## [3.1.5][] - 2023-02-03
 
 ### Added
@@ -1715,8 +1719,6 @@ _Failed released_
 [3.1.2]: https://github.com/lumapps/design-system/tree/v3.1.2
 [unreleased]: https://github.com/lumapps/design-system/compare/v3.1.3...HEAD
 [3.1.3]: https://github.com/lumapps/design-system/tree/v3.1.3
-
-
-[Unreleased]: https://github.com/lumapps/design-system/compare/v3.1.5...HEAD
+[unreleased]: https://github.com/lumapps/design-system/compare/v3.1.5...HEAD
 [3.1.5]: https://github.com/lumapps/design-system/compare/v3.1.4...v3.1.5
 [3.1.4]: https://github.com/lumapps/design-system/tree/v3.1.4